+++ src/minecraft/icbm/gangshao/render/RenderFloatingText.java
@@ -28,7 +28,7 @@ public class RenderFloatingText
 		float scale = 0.027f;
 		GL11.glColor4f(1f, 1f, 1f, 0.5f);
 		GL11.glPushMatrix();
-		GL11.glTranslatef((float) x + 0.0F, (float) y + 2.3F, (float) z);
+		GL11.glTranslatef(x + 0.0F, y + 2.3F, z);
 		GL11.glNormal3f(0.0F, 1.0F, 0.0F);
 		GL11.glRotatef(-renderManager.playerViewY, 0.0F, 1.0F, 0.0F);
 		GL11.glRotatef(renderManager.playerViewX, 1.0F, 0.0F, 0.0F);
@@ -45,10 +45,10 @@ public class RenderFloatingText
 		tessellator.startDrawingQuads();
 		int stringMiddle = fontRenderer.getStringWidth(text) / 2;
 		tessellator.setColorRGBA_F(0.0F, 0.0F, 0.0F, 0.5F);
-		tessellator.addVertex((double) (-stringMiddle - 1), (double) (-1 + yOffset), 0.0D);
-		tessellator.addVertex((double) (-stringMiddle - 1), (double) (8 + yOffset), 0.0D);
-		tessellator.addVertex((double) (stringMiddle + 1), (double) (8 + yOffset), 0.0D);
-		tessellator.addVertex((double) (stringMiddle + 1), (double) (-1 + yOffset), 0.0D);
+		tessellator.addVertex(-stringMiddle - 1, -1 + yOffset, 0.0D);
+		tessellator.addVertex(-stringMiddle - 1, 8 + yOffset, 0.0D);
+		tessellator.addVertex(stringMiddle + 1, 8 + yOffset, 0.0D);
+		tessellator.addVertex(stringMiddle + 1, -1 + yOffset, 0.0D);
 		tessellator.draw();
 		GL11.glEnable(GL11.GL_TEXTURE_2D);
 		GL11.glColor4f(1f, 1f, 1f, 0.5f);
