+++ src/minecraft/icbm/zhapin/zhapin/EShouLiuDan.java
@@ -47,16 +47,16 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 		this(par1World);
 		this.thrower = par2EntityLiving;
 		this.setSize(0.25F, 0.25F);
-		this.setLocationAndAngles(par2EntityLiving.posX, par2EntityLiving.posY + (double) par2EntityLiving.getEyeHeight(), par2EntityLiving.posZ, par2EntityLiving.rotationYaw, par2EntityLiving.rotationPitch);
-		this.posX -= (double) (MathHelper.cos(this.rotationYaw / 180.0F * (float) Math.PI) * 0.16F);
+		this.setLocationAndAngles(par2EntityLiving.posX, par2EntityLiving.posY + par2EntityLiving.getEyeHeight(), par2EntityLiving.posZ, par2EntityLiving.rotationYaw, par2EntityLiving.rotationPitch);
+		this.posX -= MathHelper.cos(this.rotationYaw / 180.0F * (float) Math.PI) * 0.16F;
 		this.posY -= 0.10000000149011612D;
-		this.posZ -= (double) (MathHelper.sin(this.rotationYaw / 180.0F * (float) Math.PI) * 0.16F);
+		this.posZ -= MathHelper.sin(this.rotationYaw / 180.0F * (float) Math.PI) * 0.16F;
 		this.setPosition(this.posX, this.posY, this.posZ);
 		this.yOffset = 0.0F;
 		float var3 = 0.4F;
-		this.motionX = (double) (-MathHelper.sin(this.rotationYaw / 180.0F * (float) Math.PI) * MathHelper.cos(this.rotationPitch / 180.0F * (float) Math.PI) * var3);
-		this.motionZ = (double) (MathHelper.cos(this.rotationYaw / 180.0F * (float) Math.PI) * MathHelper.cos(this.rotationPitch / 180.0F * (float) Math.PI) * var3);
-		this.motionY = (double) (-MathHelper.sin((this.rotationPitch) / 180.0F * (float) Math.PI) * var3);
+		this.motionX = -MathHelper.sin(this.rotationYaw / 180.0F * (float) Math.PI) * MathHelper.cos(this.rotationPitch / 180.0F * (float) Math.PI) * var3;
+		this.motionZ = MathHelper.cos(this.rotationYaw / 180.0F * (float) Math.PI) * MathHelper.cos(this.rotationPitch / 180.0F * (float) Math.PI) * var3;
+		this.motionY = -MathHelper.sin((this.rotationPitch) / 180.0F * (float) Math.PI) * var3;
 		this.setThrowableHeading(this.motionX, this.motionY, this.motionZ, 1.8f * nengLiang, 1.0F);
 		this.haoMa = explosiveID;
 	}
@@ -85,21 +85,21 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 	public void setThrowableHeading(double par1, double par3, double par5, float par7, float par8)
 	{
 		float var9 = MathHelper.sqrt_double(par1 * par1 + par3 * par3 + par5 * par5);
-		par1 /= (double) var9;
-		par3 /= (double) var9;
-		par5 /= (double) var9;
-		par1 += this.rand.nextGaussian() * 0.007499999832361937D * (double) par8;
-		par3 += this.rand.nextGaussian() * 0.007499999832361937D * (double) par8;
-		par5 += this.rand.nextGaussian() * 0.007499999832361937D * (double) par8;
-		par1 *= (double) par7;
-		par3 *= (double) par7;
-		par5 *= (double) par7;
+		par1 /= var9;
+		par3 /= var9;
+		par5 /= var9;
+		par1 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+		par3 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+		par5 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+		par1 *= par7;
+		par3 *= par7;
+		par5 *= par7;
 		this.motionX = par1;
 		this.motionY = par3;
 		this.motionZ = par5;
 		float var10 = MathHelper.sqrt_double(par1 * par1 + par5 * par5);
 		this.prevRotationYaw = this.rotationYaw = (float) (Math.atan2(par1, par5) * 180.0D / Math.PI);
-		this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, (double) var10) * 180.0D / Math.PI);
+		this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, var10) * 180.0D / Math.PI);
 	}
 
 	/**
@@ -116,7 +116,7 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 		{
 			float var7 = MathHelper.sqrt_double(par1 * par1 + par5 * par5);
 			this.prevRotationYaw = this.rotationYaw = (float) (Math.atan2(par1, par5) * 180.0D / Math.PI);
-			this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, (double) var7) * 180.0D / Math.PI);
+			this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, var7) * 180.0D / Math.PI);
 		}
 	}
 
@@ -146,10 +146,10 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 			if (ZhuYaoZhaPin.shiBaoHu(this.worldObj, new Vector3(this), ZhaPinType.SHOU_LIU_DAN, this.haoMa))
 			{
 				float var6 = 0.7F;
-				double var7 = (double) (this.worldObj.rand.nextFloat() * var6) + (double) (1.0F - var6) * 0.5D;
-				double var9 = (double) (this.worldObj.rand.nextFloat() * var6) + (double) (1.0F - var6) * 0.5D;
-				double var11 = (double) (this.worldObj.rand.nextFloat() * var6) + (double) (1.0F - var6) * 0.5D;
-				EntityItem var13 = new EntityItem(this.worldObj, (double) this.posX + var7, (double) this.posY + var9, (double) this.posZ + var11, new ItemStack(ZhuYaoZhaPin.itShouLiuDan, this.haoMa, 1));
+				double var7 = this.worldObj.rand.nextFloat() * var6 + (1.0F - var6) * 0.5D;
+				double var9 = this.worldObj.rand.nextFloat() * var6 + (1.0F - var6) * 0.5D;
+				double var11 = this.worldObj.rand.nextFloat() * var6 + (1.0F - var6) * 0.5D;
+				EntityItem var13 = new EntityItem(this.worldObj, this.posX + var7, this.posY + var9, this.posZ + var11, new ItemStack(ZhuYaoZhaPin.itShouLiuDan, this.haoMa, 1));
 				var13.delayBeforeCanPickup = 10;
 				this.worldObj.spawnEntityInWorld(var13);
 				this.setDead();
@@ -166,7 +166,7 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 		float var16 = MathHelper.sqrt_double(this.motionX * this.motionX + this.motionZ * this.motionZ);
 		this.rotationYaw = (float) (Math.atan2(this.motionX, this.motionZ) * 180.0D / Math.PI);
 
-		for (this.rotationPitch = (float) (Math.atan2(this.motionY, (double) var16) * 180.0D / Math.PI); this.rotationPitch - this.prevRotationPitch < -180.0F; this.prevRotationPitch -= 360.0F)
+		for (this.rotationPitch = (float) (Math.atan2(this.motionY, var16) * 180.0D / Math.PI); this.rotationPitch - this.prevRotationPitch < -180.0F; this.prevRotationPitch -= 360.0F)
 		{
 			;
 		}
@@ -196,15 +196,15 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 			for (int var7 = 0; var7 < 4; ++var7)
 			{
 				float var19 = 0.25F;
-				this.worldObj.spawnParticle("bubble", this.posX - this.motionX * (double) var19, this.posY - this.motionY * (double) var19, this.posZ - this.motionZ * (double) var19, this.motionX, this.motionY, this.motionZ);
+				this.worldObj.spawnParticle("bubble", this.posX - this.motionX * var19, this.posY - this.motionY * var19, this.posZ - this.motionZ * var19, this.motionX, this.motionY, this.motionZ);
 			}
 
 			var17 = 0.8F;
 		}
 
-		this.motionX *= (double) var17;
-		this.motionY *= (double) var17;
-		this.motionZ *= (double) var17;
+		this.motionX *= var17;
+		this.motionY *= var17;
+		this.motionZ *= var17;
 
 		if (this.onGround)
 		{
@@ -214,7 +214,7 @@ public class EShouLiuDan extends Entity implements IExplosiveContainer, IEntityA
 		}
 		else
 		{
-			this.motionY -= (double) gravity;
+			this.motionY -= gravity;
 			this.pushOutOfBlocks(this.posX, (this.boundingBox.minY + this.boundingBox.maxY) / 2.0D, this.posZ);
 		}
